moving EventSourceMock to __tests__ folder.
