+++ packages/shared/sdk-client/__tests__/LDClientImpl.events.test.ts
@@ -14,9 +14,9 @@ import {
 } from '@launchdarkly/private-js-mocks';
 
 import LDClientImpl from '../src/LDClientImpl';
-import { MockEventSource } from '../src/LDClientImpl.mocks';
 import { Flags } from '../src/types';
 import * as mockResponseJson from './evaluation/mockResponse.json';
+import { MockEventSource } from './streaming/LDClientImpl.mocks';
 
 type InputCustomEvent = internal.InputCustomEvent;
 type InputIdentifyEvent = internal.InputIdentifyEvent;
