+++ packages/shared/sdk-client/__tests__/LDClientImpl.timeout.test.ts
@@ -3,9 +3,9 @@ import { createBasicPlatform, createLogger } from '@launchdarkly/private-js-mock
 
 import { toMulti } from '../src/context/addAutoEnv';
 import LDClientImpl from '../src/LDClientImpl';
-import { MockEventSource } from '../src/LDClientImpl.mocks';
 import { Flags } from '../src/types';
 import * as mockResponseJson from './evaluation/mockResponse.json';
+import { MockEventSource } from './streaming/LDClientImpl.mocks';
 
 let mockPlatform: ReturnType<typeof createBasicPlatform>;
 let logger: ReturnType<typeof createLogger>;
