+++ packages/shared/sdk-client/__tests__/LDClientImpl.variation.test.ts
@@ -8,9 +8,9 @@ import {
 import { createBasicPlatform, createLogger } from '@launchdarkly/private-js-mocks';
 
 import LDClientImpl from '../src/LDClientImpl';
-import { MockEventSource } from '../src/LDClientImpl.mocks';
 import { Flags } from '../src/types';
 import * as mockResponseJson from './evaluation/mockResponse.json';
+import { MockEventSource } from './streaming/LDClientImpl.mocks';
 
 let mockPlatform: ReturnType<typeof createBasicPlatform>;
 let logger: ReturnType<typeof createLogger>;
