+++ packages/shared/sdk-client/__tests__/streaming/index.ts
@@ -1,0 +1,1 @@
+export { MockEventSource } from './LDClientImpl.mocks';
