+++ packages/shared/sdk-client/src/index.ts
@@ -18,5 +18,4 @@ export type {
 
 export { StreamingPaths } from './streaming';
 
-
 export { LDClientImpl };
